requirements.txt
flask
azure-storage-blob
azure-identity
openai
pandas
scikit-learn
shap
python-docx
openpyxl
sentence-transformers
fastapi
uvicorn
joblib
json
os
re
requests
azure-ai-agents
azure-ai-projects
requests
gunicorn